release: 0.0.1
Finish the most basic PoC

It gets the current HEAD, goes through the history until it finds a
a commit that starts with "release:", then parses said version
according to SemVer.

From then on, it reads the commits from the HEAD, until said release
commit, and looks for commits with a title starting with "fix", "feat"
or "feat!" for patch, minor and major bumps, respectively.